// src/app/core/models/register-customer.model.ts
export class RegisterCustomerModel {
    emailAddress: string;
    password: string;
    firstName: string;
    lastName: string;
    primaryPhone: string;
    alternatePhone: string;
    streetAddress: string;
    city: string;
    state: string;
    zipCode: string;
    country: number;
    aboutUs: string;
    refererEmail: string;
    callingCountryId: number;
    isQuickSignUp: boolean;
    PromoCode:string;
}